Fixing header includes and size_type

+++ dev/plugins/hu.elte.txtuml.export.cpp/cpp-runtime/itimer.cpp
@@ -1,6 +1,6 @@
 #include "itimer.hpp"
 #include "timer.hpp"
 
-Timer *ITimer::start(StateMachineI *sm, EventPtr event, long millisecs) {
+Timer *ITimer::start(StateMachineI *sm, EventPtr event, int millisecs) {
     return new Timer(sm,event,millisecs);
 }
